Extract types to separate files
